Add files via upload
The chi-square value of a matrix m x n (observed values) is compared with the chi-square values of N permutations of the same matrix (raw and column sums kept constant). p-value is obtained by the rate between the number of times the permuted matrices have higher chi-square values and the total number of permutations. Standard error of p-value is calculated as well.